Added ability to organise MP4 videos

+++ src/test/java/com/benjaminsproule/photoorganiser/util/MimeTypesUtilITest.java
@@ -1,6 +1,5 @@
 package com.benjaminsproule.photoorganiser.util;
 
-import org.hamcrest.Matchers;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
@@ -10,10 +9,10 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+import static com.benjaminsproule.photoorganiser.util.MimeTypesUtil.*;
 import static java.util.Arrays.asList;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.hasItem;
-import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.*;
 
 public class MimeTypesUtilITest {
 
@@ -22,6 +21,10 @@ public class MimeTypesUtilITest {
 
     @Before
     public void setup() throws Exception {
+        if (!MimeTypesUtil.requiresMimeTypesFile()) {
+            return;
+        }
+
         String home = System.getProperty("user.home");
         mimeTypes = new File(home + "/.mime.types").toPath();
         mimeTypesBackup = new File(home + "/.mime.types.backup").toPath();
@@ -34,6 +37,10 @@ public class MimeTypesUtilITest {
 
     @After
     public void tearDown() throws Exception {
+        if (!MimeTypesUtil.requiresMimeTypesFile()) {
+            return;
+        }
+
         Files.delete(mimeTypes);
 
         if (Files.exists(mimeTypesBackup)) {
@@ -63,7 +70,7 @@ public class MimeTypesUtilITest {
 
         MimeTypesUtil.createMimeTypesFile();
 
-        assertThat(Files.readAllLines(mimeTypes), Matchers.hasItems(MimeTypesUtil.IMAGES_JPG_MAPPING, MimeTypesUtil.IMAGES_TIFF));
+        assertThat(Files.readAllLines(mimeTypes), hasItems(IMAGES_JPG_MAPPING, IMAGES_TIFF, VIDEOS_MP4));
     }
 
     @Test
@@ -85,11 +92,11 @@ public class MimeTypesUtilITest {
             return;
         }
 
-        populateMimeTypesFile(MimeTypesUtil.IMAGES_JPG_MAPPING, MimeTypesUtil.IMAGES_TIFF);
+        populateMimeTypesFile(IMAGES_JPG_MAPPING, IMAGES_TIFF, VIDEOS_MP4);
 
         MimeTypesUtil.createMimeTypesFile();
 
-        assertThat(Files.readAllLines(mimeTypes), Matchers.containsInAnyOrder(MimeTypesUtil.IMAGES_JPG_MAPPING, MimeTypesUtil.IMAGES_TIFF));
+        assertThat(Files.readAllLines(mimeTypes), containsInAnyOrder(IMAGES_JPG_MAPPING, IMAGES_TIFF, VIDEOS_MP4));
     }
 
     private void populateMimeTypesFile(final String... str) throws IOException {
